backend/requirements.txt
# Core dependencies
google-generativeai>=0.3.2
requests>=2.31.0
beautifulsoup4>=4.12.2
playwright>=1.40.0
python-dotenv>=1.0.0
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
jinja2>=3.1.2
python-multipart>=0.0.6
aiofiles>=23.2.1
pandas>=1.3.0

# File processing
lxml>=4.9.3
openpyxl>=3.1.2
python-docx>=1.1.0
PyPDF2>=3.0.1

# Job search
python-jobspy>=1.1.79

# Database dependencies
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
aiosqlite>=0.19.0

# Development dependencies (optional)
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.0.0 